New EdgeOnlyImageProcessorFragment for Pose, Object, and Triton to inherit from
