+++ android/MobiledgeXSDKDemo/computervision/src/main/java/com/mobiledgex/computervision/PoseProcessorFragment.java
@@ -46,7 +46,7 @@ import org.json.JSONArray;
 
 import static com.mobiledgex.matchingenginehelper.MatchingEngineHelper.DEF_HOSTNAME_PLACEHOLDER;
 
-public class PoseProcessorFragment extends ImageProcessorFragment implements ImageServerInterface,
+public class PoseProcessorFragment extends EdgeOnlyImageProcessorFragment implements ImageServerInterface,
         ImageProviderInterface {
     private static final String TAG = "PoseProcessorFragment";
     public static final String EXTRA_POSE_STROKE_WIDTH = "EXTRA_POSE_STROKE_WIDTH";
@@ -56,40 +56,6 @@ public class PoseProcessorFragment extends ImageProcessorFragment implements Ima
 
     public static PoseProcessorFragment newInstance() {
         return new PoseProcessorFragment();
-    }
-
-    public String getStatsText() {
-        if (mImageSenderEdge != null) {
-            return mImageSenderEdge.getStatsText();
-        } else {
-            return "Edge never initialized.";
-        }
-    }
-
-    /**
-     * Perform any processing of the given bitmap.
-     *
-     * @param bitmap  The bitmap from the camera or video.
-     * @param imageRect  The coordinates of the image on the screen. Needed for scaling/offsetting
-     *                   resulting pose skeleton coordinates.
-     */
-    @Override
-    public void onBitmapAvailable(Bitmap bitmap, Rect imageRect) {
-        if(bitmap == null) {
-            return;
-        }
-
-        mImageRect = imageRect;
-        mServerToDisplayRatioX = (float) mImageRect.width() / bitmap.getWidth();
-        mServerToDisplayRatioY = (float) mImageRect.height() / bitmap.getHeight();
-
-        Log.d(TAG, "mImageRect="+mImageRect.toShortString()+" mImageRect.height()="+mImageRect.height()+" bitmap.getWidth()="+bitmap.getWidth()+" bitmap.getHeight()="+bitmap.getHeight()+" mServerToDisplayRatioX=" + mServerToDisplayRatioX +" mServerToDisplayRatioY=" + mServerToDisplayRatioY);
-
-        if (mImageSenderEdge != null) {
-            mImageSenderEdge.sendImage(bitmap);
-        } else {
-            Log.d(TAG, "Waiting for mImageSenderEdge to be initialized.");
-        }
     }
 
     /**
@@ -129,11 +95,6 @@ public class PoseProcessorFragment extends ImageProcessorFragment implements Ima
                 mPoseRenderer.invalidate();
             }
         });
-    }
-
-    @Override
-    public void updateTrainingProgress(int trainingCount, ImageSender.CameraMode mode) {
-        throw new UnsupportedOperationException();
     }
 
     @Override
@@ -202,91 +163,6 @@ public class PoseProcessorFragment extends ImageProcessorFragment implements Ima
             restartImageSenderEdge();
         } else {
             meHelper.findCloudletInBackground();
-        }
-    }
-
-    @Override
-    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
-        Log.i(TAG, "onCreateOptionsMenu");
-        mOptionsMenu = menu;
-        // Inflate the menu; this adds items to the action bar if it is present.
-        inflater.inflate(R.menu.camera_menu, menu);
-
-        //Remove these menu items.
-        menu.findItem(R.id.action_camera_training).setVisible(false);
-        menu.findItem(R.id.action_camera_training_guest).setVisible(false);
-        menu.findItem(R.id.action_camera_remove_training_data).setVisible(false);
-        menu.findItem(R.id.action_camera_remove_training_guest_data).setVisible(false);
-
-        //No Cloud available for benchmarking
-        menu.findItem(R.id.action_benchmark_cloud).setVisible(false);
-
-        // Declutter the menu, but keep the code in place in case we need it later.
-        menu.findItem(R.id.action_camera_debug).setVisible(false);
-    }
-
-    @Override
-    protected void toggleViews() {
-        if(prefShowFullLatency) {
-            mEdgeLatencyFull.setVisibility(View.VISIBLE);
-        } else {
-            mEdgeLatencyFull.setVisibility(View.INVISIBLE);
-            mEdgeStdFull.setVisibility(View.GONE);
-            mEdgeStdNet.setVisibility(View.GONE);
-        }
-        if(prefShowNetLatency) {
-            mEdgeLatencyNet.setVisibility(View.VISIBLE);
-            mEdgeStdNet.setVisibility(View.VISIBLE);
-        } else {
-            mEdgeLatencyNet.setVisibility(View.INVISIBLE);
-            mEdgeStdNet.setVisibility(View.GONE);
-        }
-        if(prefShowStdDev) {
-            if(prefShowNetLatency) {
-                mEdgeStdFull.setVisibility(View.VISIBLE);
-            }
-            if(prefShowNetLatency) {
-                mEdgeStdNet.setVisibility(View.VISIBLE);
-            }
-        } else {
-            mEdgeStdFull.setVisibility(View.GONE);
-            mEdgeStdNet.setVisibility(View.GONE);
-        }
-    }
-
-    @Override
-    public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key) {
-        Log.i(TAG, "Pose Detection onSharedPreferenceChanged("+key+")");
-        if(getContext() == null) {
-            //Can happen during rapid screen rotations.
-            return;
-        }
-        super.onSharedPreferenceChanged(sharedPreferences, key);
-
-        String prefKeyHostGpuOverride = getResources().getString(R.string.pref_override_gpu_cloudlet_hostname);
-        String prefKeyHostGpu = getResources().getString(R.string.preference_gpu_host_edge);
-
-        if (key.equals(prefKeyHostGpuOverride) || key.equals("ALL")) {
-            mGpuHostNameOverride = sharedPreferences.getBoolean(prefKeyHostGpuOverride, false);
-            Log.i(TAG, "key="+key+" mGpuHostNameOverride="+ mGpuHostNameOverride);
-            if (mGpuHostNameOverride) {
-                mHostDetectionEdge = sharedPreferences.getString(prefKeyHostGpu, DEF_HOSTNAME_PLACEHOLDER);
-                Log.i(TAG, "key="+key+" mHostDetectionEdge="+ mHostDetectionEdge);
-            }
-        }
-        if (key.equals(prefKeyHostGpu) || key.equals("ALL")) {
-            mHostDetectionEdge = sharedPreferences.getString(prefKeyHostGpu, DEF_HOSTNAME_PLACEHOLDER);
-            Log.i(TAG, "key="+key+" mHostDetectionEdge="+ mHostDetectionEdge);
-        }
-
-        if (key.equals(prefKeyHostGpu) || key.equals(prefKeyHostGpuOverride)) {
-            if (mGpuHostNameOverride) {
-                mEdgeHostList.clear();
-                mEdgeHostListIndex = 0;
-                mEdgeHostList.add(mHostDetectionEdge);
-                showMessage("mHostDetectionEdge set to " + mHostDetectionEdge);
-                restartImageSenderEdge();
-            }
         }
     }
 }
