+++ android/MobiledgeXSDKDemo/computervision/src/main/res/layout/settings_activity.xml
@@ -7,8 +7,4 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 
-    <FrameLayout
-        android:id="@+id/settings"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
 </LinearLayout>
