+++ android/MobiledgeXSDKDemo/matchingenginehelper/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
     <application>
         <activity
             android:name=".SettingsActivity"
-            android:label="@string/title_activity_settings_x"></activity>
+            android:label="@string/title_activity_settings"></activity>
     </application>
 
 </manifest>
